update modal in case info and case file tab
